This restores the movement by dragging
The function to move by dragging was broken. This restores it.

It also makes the movement on click take effect when the mouse button is
released, which is the common way to react to clicks.

I have cleaned up the previous code for drag, which didn't appear to work.

It is needed to make the game playable on touchscreens.

I have tested it myself on my mobian device and on my desktop device.